Implementing responses using status code HTTP

+++ app/Http/Controllers/MarcaController.php
@@ -7,13 +7,20 @@ use Illuminate\Http\Request;
 
 class MarcaController extends Controller
 {
+    protected $marca;
+
+    public function __construct(Marca $marca)
+    {
+        $this->marca = $marca;
+    }
+
     /**
      * Display a listing of the resource.
      */
     public function index()
     {
-        $results = Marca::all();
-        return $results;
+        $results = $this->marca->all();
+        return response()->json($results, 200);
     }
 
     /**
@@ -21,33 +28,51 @@ class MarcaController extends Controller
      */
     public function store(Request $request)
     {
-        $result = Marca::create($request->all());
-        return $result;
+        $result = $this->marca->create($request->all());
+        return response()->json($result, 201);
     }
 
     /**
      * Display the specified resource.
      */
-    public function show(Marca $marca)
+    public function show($id)
     {
-        return $marca;
+        $marca = $this->marca->find($id);
+
+        if ($marca === null) {
+            return response()->json(['erro' => 'Recurso pesquisado não existe'], 404);
+        }
+
+        return response()->json($marca, 200);
     }
 
     /**
      * Update the specified resource in storage.
      */
-    public function update(Request $request, Marca $marca)
+    public function update(Request $request, $id)
     {
+        $marca = $this->marca->find($id);
+
+        if ($marca === null) {
+            return response()->json(['erro' => 'Impossível realizar a atualização: o recurso solicitado não existe.'], 404);
+        }
+
         $marca->update($request->all());
-        return $marca;
+        return response()->json($marca, 200);
     }
 
     /**
      * Remove the specified resource from storage.
      */
-    public function destroy(Marca $marca)
+    public function destroy($id)
     {
+        $marca = $this->marca->find($id);
+
+        if ($marca === null) {
+            return response()->json(['erro' => 'Impossível realizar a exclusão: o recurso solicitado não existe.'], 404);
+        }
+
         $marca->delete();
-        return ['msg' => 'Marca removida com sucesso!'];
+        return response()->json(['msg' => 'Marca removida com sucesso!'], 200);
     }
 }
